+++ src/main/java/com/jrealm/data/config/WebMvcConfig.java
@@ -3,19 +3,22 @@ package com.jrealm.data.config;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.jrealm.data.service.PrivelegedAccessFilter;
+import com.jrealm.data.auth.PrivelegedAccessFilter;
 
 @Configuration
-public class WebMvcConfig extends WebMvcConfigurerAdapter {
+public class WebMvcConfig implements WebMvcConfigurer {
 
-  @Autowired 
-  private PrivelegedAccessFilter yourInjectedInterceptor;
+	@Autowired
+	private PrivelegedAccessFilter accessFilter;
+	
+	public WebMvcConfig(@Autowired final PrivelegedAccessFilter accessFilter) {
+		this.accessFilter = accessFilter;
+	}
 
-  @Override
-  public void addInterceptors(InterceptorRegistry registry) {
-	  	registry.addInterceptor(this.yourInjectedInterceptor)
-	  	.addPathPatterns("/data/**", "/admin/**");
-  }
+	@Override
+	public void addInterceptors(InterceptorRegistry registry) {
+		registry.addInterceptor(this.accessFilter).addPathPatterns("/data/**", "/admin/**");
+	}
 }
